Added capture_delay config to set input FPS
Existing code had a hard coded sleep of 5s between stills grabbed from the camera.  This effectively limited the video (regardless of the frame rate that was set) to only represent 12 frames per minute of what the camera sees.  This parameter replaces the hardcoded sleep with a configurable value, which can be made less than 1.  While the fastest frame rate achievable may be limited by other factors, this allows for smoother source video to be generated from the stills.

We could use the frame rate to calculate the value (perhaps that was the original intent?) but this allows the output frame rate of ffmpeg to be varied separately from the capture rate.